Refactor List component into its own file
